Move SpanLink and SpanEvent behavior to patches
This more clearly scopes the behavior to OpenTelemetry-only code.

The harvest! method is left as it is so that it is consistent in all
cases, even though the span array will not have four elements unless
SpanLink or SpanEvent events are present.
