+++ lib/new_relic/agent/opentelemetry/transaction_patch.rb
@@ -31,6 +31,26 @@ module NewRelic
           super
         end
 
+        def add_span_link(link)
+          initial_segment&.add_span_link(link)
+        end
+
+        def span_links
+          initial_segment&.span_links || NewRelic::EMPTY_ARRAY
+        end
+
+        # This method adds a SpanEvent event to the Transaction's initial segment.
+        # A SpanEvent is used to denote a meaningful, singular point in a
+        # Span's duration.
+        def add_span_event(name, attributes: nil, timestamp: nil)
+          initial_segment&.add_span_event(name, attributes: attributes, timestamp: timestamp)
+        end
+
+        # Used to reference SpanEvent events associated with the initial segment.
+        def span_events
+          initial_segment&.span_events || NewRelic::EMPTY_ARRAY
+        end
+
         private
 
         def find_or_create_span(segment)
