+++ lib/new_relic/agent/transaction/abstract_segment.rb
@@ -26,10 +26,6 @@ module NewRelic
 
         CALLBACK = :@callback
         SEGMENT = 'segment'
-        MAX_SPAN_LINKS = 100
-        SPAN_LINKS_DROPPED_METRIC = 'Supportability/Ruby/SpanEvent/Links/Dropped'
-        MAX_SPAN_EVENTS = 100
-        SPAN_EVENTS_DROPPED_METRIC = 'Supportability/Ruby/SpanEvent/Events/Dropped'
 
         def initialize(name = nil, start_time = nil)
           @name = name
@@ -99,48 +95,6 @@ module NewRelic
           force_finish unless finished?
           record_exclusive_duration
           record_metrics if record_metrics?
-        end
-
-        def add_span_link(link)
-          @span_links ||= []
-          if @span_links.size >= MAX_SPAN_LINKS
-            NewRelic::Agent.record_metric(SPAN_LINKS_DROPPED_METRIC, 1)
-            return
-          end
-          @span_links << link
-        end
-
-        def span_links
-          instance_variable_defined?(:@span_links) ? @span_links : NewRelic::EMPTY_ARRAY
-        end
-
-        # This method adds SpanEvent events to a Span event/Segment.
-        # A SpanEvent is used to denote a meaningful, singular point in a
-        # Span's duration.
-        def add_span_event(name, attributes: nil, timestamp: nil)
-          @span_events ||= []
-
-          if @span_events.size >= MAX_SPAN_EVENTS
-            NewRelic::Agent.record_metric(SPAN_EVENTS_DROPPED_METRIC, 1)
-            return
-          end
-
-          # Normalize to Float seconds at storage time.
-          # OTel may pass Integer nanoseconds; New Relic uses Float seconds.
-          ts = if timestamp.nil?
-            Process.clock_gettime(Process::CLOCK_REALTIME)
-          elsif timestamp.is_a?(Integer)
-            timestamp / 1_000_000_000.0
-          else
-            timestamp
-          end
-
-          @span_events << {name: name, attributes: attributes, timestamp: ts}
-        end
-
-        # Used to reference SpanEvent events associated with a Span/Segment
-        def span_events
-          instance_variable_defined?(:@span_events) ? @span_events : NewRelic::EMPTY_ARRAY
         end
 
         def params
